change readme.md and refactor type
